transport: widen out_list_s offset/length to uint64_t for >4 GiB resends
partial_send_store accumulates byte offsets that overflow a 32-bit field for
transfers larger than 4 GiB, silently corrupting the resend. offsetPayload was
already 64-bit; this widens the remaining offset/length on the same path and
the async_free offset parameter.

+++ libs/include/internal/arts/transport/protocol.h
@@ -252,7 +252,7 @@ void arts_remote_send_request_payload_async(int rank, char *message,
                                             uint64_t size);
 void arts_remote_send_request_payload_async_free(
     int rank, char *message, unsigned int length, char *payload,
-    unsigned int offset, uint64_t size, void (*free_method)(void *));
+    uint64_t offset, uint64_t size, void (*free_method)(void *));
 void arts_remote_set_thread_outbound_queues(unsigned int start,
                                             unsigned int stop);
 void arts_remote_thread_outbound_queues_cleanup();
